Set minTimestamp correctly to be able to drop expired sstables correctly
patch by marcuse, reviewed by jbellis for CASSANDRA-6337

+++ src/java/org/apache/cassandra/db/compaction/CompactionController.java
@@ -106,7 +106,7 @@ public class CompactionController
 
         List<SSTableReader> candidates = new ArrayList<SSTableReader>();
 
-        long minTimestamp = Integer.MAX_VALUE;
+        long minTimestamp = Long.MAX_VALUE;
 
         for (SSTableReader sstable : overlapping)
             minTimestamp = Math.min(minTimestamp, sstable.getMinTimestamp());
